+++ packages/auto-approve/src/auto-approve.ts
@@ -142,7 +142,7 @@ export function handler(app: Probot) {
     ],
     async (context: Context) => {
       const pr = context.payload;
-      const owner = pr.pull_request.head.repo.owner.login;
+      const owner = pr.repository.owner.login;
       const repo = pr.pull_request.head.repo.name;
       const prNumber = pr.number;
 
@@ -195,7 +195,7 @@ export function handler(app: Probot) {
 
         logger.metric('auto_approve.status_check', {
           repo: `${owner}/${repo}`,
-          pr: pr,
+          pr: prNumber,
         });
       } else {
         let config: Configuration | null;
@@ -257,7 +257,7 @@ export function handler(app: Probot) {
             });
             logger.metric('auto_approve.approved_tagged', {
               repo: `${owner}/${repo}`,
-              pr: pr,
+              pr: prNumber,
             });
             logger.info(
               `Auto-approved and tagged ${owner}/${repo}/${prNumber}`