+++ packages/auto-label/src/auto-label.ts
@@ -35,8 +35,6 @@ import {
 
 type IssueResponse = Endpoints['GET /repos/{owner}/{repo}/issues']['response'];
 
-import colorsData from './colors.json';
-
 const storage = new Storage();
 
 handler.getDriftFile = async (file: string) => {
@@ -94,26 +92,9 @@ handler.addLabeltoRepoAndIssue = async function addLabeltoRepoAndIssue(
     const apis = await handler.getDriftApis();
     autoDetectedLabel = helper.autoDetectLabel(apis, issueTitle);
   }
-  const index = driftRepos?.findIndex(r => driftRepo === r) % colorsData.length;
-  const colorNumber = index >= 0 ? index : 0;
   const githubLabel = driftRepo?.github_label || autoDetectedLabel;
 
   if (githubLabel) {
-    try {
-      await context.octokit.issues.createLabel({
-        owner,
-        repo,
-        name: githubLabel,
-        color: colorsData[colorNumber].color,
-      });
-      logger.info(`Label added to ${owner}/${repo} is ${githubLabel}`);
-    } catch (e) {
-      // HTTP 422 means the label already exists on the repo
-      if (e.status !== 422) {
-        e.message = `Error creating label: ${e.message}`;
-        logger.error(e);
-      }
-    }
     if (labelsOnIssue) {
       const foundAPIName = helper.labelExists(labelsOnIssue, githubLabel);
 
@@ -170,14 +151,6 @@ handler.addLabeltoRepoAndIssue = async function addLabeltoRepoAndIssue(
   const isSampleIssue =
     repo.includes('samples') || issueTitle?.includes('sample');
   if (!foundSamplesTag && isSampleIssue) {
-    await context.octokit.issues
-      .createLabel({
-        owner,
-        repo,
-        name: 'samples',
-        color: colorsData[colorNumber].color,
-      })
-      .catch(logger.error);
     await context.octokit.issues
       .addLabels({
         owner,