+++ scripts/publish-function.sh
@@ -50,16 +50,26 @@ pushd "${targetDir}"
 functionName=${botName//-/_}
 queueName=${botName//_/-}
 
+deployArgs=(
+  "--trigger-http"
+  "--runtime"
+  "${functionRuntime}"
+  "--region"
+  "${functionRegion}"
+  "--update-env-vars"
+  "DRIFT_PRO_BUCKET=${bucket},KEY_LOCATION=${keyLocation},KEY_RING=${keyRing},GCF_SHORT_FUNCTION_NAME=${functionName},PROJECT_ID=${project},GCF_LOCATION=${functionRegion},PUPPETEER_SKIP_CHROMIUM_DOWNLOAD='1',WEBHOOK_TMP=tmp-webhook-payloads"
+  "--timeout"
+  "${timeout}"
+)
+if [ -n "${SERVICE_ACCOUNT}" ]; then
+  deployArgs+=( "--service-account" "${SERVICE_ACCOUNT}" )
+fi
 echo "About to publish function ${functionName}"
-gcloud functions deploy "${functionName}" \
-  --trigger-http \
-  --runtime "${functionRuntime}" \
-  --region "${functionRegion}" \
-  --update-env-vars DRIFT_PRO_BUCKET="${bucket}",KEY_LOCATION="${keyLocation}",KEY_RING="${keyRing}",GCF_SHORT_FUNCTION_NAME="${functionName}",PROJECT_ID="${project}",GCF_LOCATION="${functionRegion}",PUPPETEER_SKIP_CHROMIUM_DOWNLOAD='1',WEBHOOK_TMP=tmp-webhook-payloads \
-  --timeout "${timeout}"
+echo gcloud functions deploy "${functionName}" "${deployArgs[@]}"
+gcloud functions deploy "${functionName}" "${deployArgs[@]}"
 
 echo "Adding ability for allUsers to execute the Function"
-gcloud alpha functions add-iam-policy-binding "${functionName}" \
+gcloud functions add-iam-policy-binding "${functionName}" \
   --region="${functionRegion}" \
   --member=allUsers \
   --role=roles/cloudfunctions.invoker